Refactor YAML code
